% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_xvg}
\alias{plot_xvg}
\title{plot xvg data}
\usage{
plot_xvg(xvg_data, title = NULL, subtitle = NULL, ...)
}
\arguments{
\item{xvg_data}{xvg data object returned by read_xvg}

\item{title}{chart title (default uses xvg file's title)}

\item{subtitle}{chart subtitle (default uses xvg file's subtitle)}

\item{...}{additional parameters passed to ggplot2::geom_line}
}
\value{
a ggplot2 object
}
\description{
plot xvg data using ggplot2
}
