% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{read_xvg}
\alias{read_xvg}
\title{read xvg files}
\usage{
read_xvg(xvg_files, skip_comments = TRUE)
}
\arguments{
\item{xvg_files}{character vector of xvg file paths}

\item{skip_comments}{logical indicating whether to skip comment lines (default: TRUE)}
}
\value{
Named list containing xvg data, using filenames (without extension) as keys
}
\description{
read one or more GROMACS-generated xvg files
}
